requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
peewee==3.17.8
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-dotenv==1.0.1
stripe==11.4.1
boto3==1.35.0
